9 backend moderators

+++ formula1-tournament-api/Controllers/DriverController.cs
@@ -11,17 +11,19 @@ namespace formula1_tournament_api.Controllers
     public class DriverController : Controller
     {
         private IDriver _driverService;
+        private IUserSeason _userSeasonService;
 
-        public DriverController(IDriver driverService)
+        public DriverController(IDriver driverService, IUserSeason userSeasonService)
         {
             _driverService = driverService;
+            _userSeasonService = userSeasonService;
         }
 
-        [HttpGet]
+        [HttpGet("{seasonId}")]
         [EnableQuery]
-        public async Task<IActionResult> Get()
+        public async Task<IActionResult> Get(Guid seasonId)
         {
-            var result = await _driverService.GetAllDrivers();
+            var result = await _driverService.GetAllDriversBySeasonId(seasonId);
             if (result.IsSuccess)
             {
                 return Ok(result.Drivers);
@@ -29,8 +31,8 @@ namespace formula1_tournament_api.Controllers
             return NotFound(result.ErrorMessage);
         }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> Get(Guid id)
+        [HttpGet("{seasonId}/{id}")]
+        public async Task<IActionResult> Get(Guid seasonId, Guid id)
         {
             var result = await _driverService.GetDriverById(id);
             if (result.IsSuccess)
@@ -40,11 +42,13 @@ namespace formula1_tournament_api.Controllers
             return NotFound(result.ErrorMessage);
         }
 
-        [HttpPost]
+        [HttpPost("{seasonId}")]
         [Authorize]
-        public async Task<IActionResult> Post([FromBody] Driver driver)
+        public async Task<IActionResult> Post(Guid seasonId, [FromBody] Driver driver)
         {
-            var result = await _driverService.AddDriver(driver, new Guid(User.Identity.Name));
+            if (!_userSeasonService.HasPermission(new Guid(User.Identity.Name), seasonId))
+                return StatusCode(StatusCodes.Status403Forbidden);
+            var result = await _driverService.AddDriver(driver);
             if (result.IsSuccess)
             {
                 return StatusCode(StatusCodes.Status201Created);
@@ -52,9 +56,11 @@ namespace formula1_tournament_api.Controllers
             return BadRequest(result.ErrorMessage);
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> Put(Guid id, [FromBody] Driver driver)
+        [HttpPut("{seasonId}/{id}")]
+        public async Task<IActionResult> Put(Guid seasonId, Guid id, [FromBody] Driver driver)
         {
+            if (!_userSeasonService.HasPermission(new Guid(User.Identity.Name), seasonId))
+                return StatusCode(StatusCodes.Status403Forbidden);
             var result = await _driverService.UpdateDriver(id, driver);
             if (result.IsSuccess)
             {
@@ -63,9 +69,11 @@ namespace formula1_tournament_api.Controllers
             return BadRequest(result.ErrorMessage);
         }
 
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(Guid id)
+        [HttpDelete("{seasonId}/{id}")]
+        public async Task<IActionResult> Delete(Guid seasonId, Guid id)
         {
+            if (!_userSeasonService.HasPermission(new Guid(User.Identity.Name), seasonId))
+                return StatusCode(StatusCodes.Status403Forbidden);
             var result = await _driverService.DeleteDriver(id);
             if (result.IsSuccess)
             {
